--- NugetAnalyzer.Web/Models/GitHubUserClaimTypes.cs
@@ -1,13 +1,0 @@
-﻿namespace NugetAnalyzer.Web.Models
-{
-    /// <summary>
-    /// Defines constants for GitHub user information keys.
-    /// </summary>
-    public class GitHubUserClaimTypes
-    {
-        public const string Id = "id";
-        public const string UserName = "login";
-        public const string Url = "html_url";
-        public const string AvatarUrl = "avatar_url";
-    }
-}
