--- NugetAnalyzer.Web/Models/OAuthSourceNames.cs
@@ -1,7 +1,0 @@
-﻿namespace NugetAnalyzer.Web.Models
-{
-    public static class OAuthSourceNames
-    {
-        public const string GitHubSourceName = "GitHub";
-    }
-}
